Ajuste

+++ src/main/java/br/com/crechesystem/crechesystem/service/TurmaServiceImpl.java
@@ -36,7 +36,7 @@ public class TurmaServiceImpl implements TurmaService {
     }
 
     @Override
-    public Optional<Turma> findByBirthMonth(long months) {
+    public Optional<Turma> findByBirthMonth(int months) {
         return this.turmaRepository.findTop1ByMinimumMonthsLessThanEqualByMaximumMonthsLessThanEqual(months);
     }
 }
